Added with_cache() and refactored basic element-wise transforms (#2443)
