+++ pyro/distributions/transforms/__init__.py
@@ -13,6 +13,7 @@ from pyro.distributions.transforms.affine_autoregressive import (AffineAutoregre
 from pyro.distributions.transforms.affine_coupling import (AffineCoupling, ConditionalAffineCoupling, affine_coupling,
                                                            conditional_affine_coupling)
 from pyro.distributions.transforms.batchnorm import BatchNorm, batchnorm
+from pyro.distributions.transforms.basic import ELUTransform, LeakyReLUTransform, elu, leaky_relu
 from pyro.distributions.transforms.block_autoregressive import BlockAutoregressive, block_autoregressive
 from pyro.distributions.transforms.cholesky import CorrLCholeskyTransform
 from pyro.distributions.transforms.discrete_cosine import DiscreteCosineTransform
@@ -23,10 +24,11 @@ from pyro.distributions.transforms.generalized_channel_permute import (Condition
 from pyro.distributions.transforms.householder import (ConditionalHouseholder, Householder, conditional_householder,
                                                        householder)
 from pyro.distributions.transforms.lower_cholesky_affine import LowerCholeskyAffine
-from pyro.distributions.transforms.neural_autoregressive import (ConditionalNeuralAutoregressive, ELUTransform,
-                                                                 LeakyReLUTransform, NeuralAutoregressive,
-                                                                 conditional_neural_autoregressive, elu, leaky_relu,
-                                                                 neural_autoregressive, tanh)
+from pyro.distributions.transforms.neural_autoregressive import (
+    ConditionalNeuralAutoregressive,
+    NeuralAutoregressive,
+    conditional_neural_autoregressive,
+    neural_autoregressive)
 from pyro.distributions.transforms.permute import Permute, permute
 from pyro.distributions.transforms.planar import ConditionalPlanar, Planar, conditional_planar, planar
 from pyro.distributions.transforms.polynomial import Polynomial, polynomial
@@ -111,8 +113,7 @@ __all__ = [
     'polynomial',
     'radial',
     'spline',
-    'sylvester',
-    'tanh',
+    'sylvester'
 ]
 
 __all__.extend(torch_transforms)
