+++ pyro/distributions/transforms/neural_autoregressive.py
@@ -1,94 +1,21 @@
 # Copyright (c) 2017-2019 Uber Technologies, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-from __future__ import absolute_import, division, print_function
-
-import math
 from functools import partial
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.distributions import constraints
-from torch.distributions.transforms import SigmoidTransform, TanhTransform, Transform
+from torch.distributions.transforms import SigmoidTransform, TanhTransform
 
 from pyro.distributions.conditional import ConditionalTransformModule
 from pyro.distributions.torch_transform import TransformModule
+from pyro.distributions.transforms.basic import ELUTransform, LeakyReLUTransform
 from pyro.distributions.util import copy_docs_from
 from pyro.nn import AutoRegressiveNN, ConditionalAutoRegressiveNN
 
 eps = 1e-8
-
-
-class ELUTransform(Transform):
-    r"""
-    Bijective transform via the mapping :math:`y = \text{ELU}(x)`.
-    """
-    domain = constraints.real
-    codomain = constraints.positive
-    bijective = True
-    sign = +1
-
-    def __eq__(self, other):
-        return isinstance(other, ELUTransform)
-
-    def _call(self, x):
-        return F.elu(x)
-
-    def _inverse(self, y):
-        return torch.max(y, torch.zeros_like(y)) + torch.min(torch.log1p(y + eps), torch.zeros_like(y))
-
-    def log_abs_det_jacobian(self, x, y):
-        return -F.relu(-x)
-
-
-def elu():
-    """
-    A helper function to create an
-    :class:`~pyro.distributions.transform.ELUTransform` object for consistency with
-    other helpers.
-    """
-    return ELUTransform()
-
-
-class LeakyReLUTransform(Transform):
-    r"""
-    Bijective transform via the mapping :math:`y = \text{LeakyReLU}(x)`.
-    """
-    domain = constraints.real
-    codomain = constraints.positive
-    bijective = True
-    sign = +1
-
-    def __eq__(self, other):
-        return isinstance(other, LeakyReLUTransform)
-
-    def _call(self, x):
-        return F.leaky_relu(x)
-
-    def _inverse(self, y):
-        return F.leaky_relu(y, negative_slope=100.0)
-
-    def log_abs_det_jacobian(self, x, y):
-        return torch.where(x >= 0., torch.zeros_like(x), torch.ones_like(x) * math.log(0.01))
-
-
-def leaky_relu():
-    """
-    A helper function to create a
-    :class:`~pyro.distributions.transforms.LeakyReLUTransform` object for
-    consistency with other helpers.
-    """
-    return LeakyReLUTransform()
-
-
-def tanh():
-    """
-    A helper function to create a
-    :class:`~pyro.distributions.transforms.TanhTransform` object for consistency
-    with other helpers.
-    """
-    return TanhTransform()
 
 
 @copy_docs_from(TransformModule)
@@ -135,6 +62,7 @@ class NeuralAutoregressive(TransformModule):
     codomain = constraints.real
     bijective = True
     event_dim = 1
+    eps = 1e-8
     autoregressive = True
 
     def __init__(self, autoregressive_nn, hidden_units=16, activation='sigmoid'):
@@ -201,7 +129,7 @@ class NeuralAutoregressive(TransformModule):
         T = self.T
 
         log_dydD = self._cached_log_df_inv_dx
-        log_dDdx = torch.logsumexp(torch.log(A + eps) + self.logsoftmax(W_pre) +
+        log_dDdx = torch.logsumexp(torch.log(A + self.eps) + self.logsoftmax(W_pre) +
                                    T.log_abs_det_jacobian(C, T_C), dim=-2)
         log_det = log_dydD + log_dDdx
         return log_det.sum(-1)
